+++ common/container/CMakeLists.txt
@@ -2,4 +2,3 @@ project(cpp_common_container)
 
 aux_source_directory(. SRCS)
 
-add_library(${PROJECT_NAME} ${SRCS})
