+++ common/test/CMakeLists.txt
@@ -5,7 +5,6 @@ aux_source_directory(. TEST_SRCS)
 add_executable(${PROJECT_NAME} ${TEST_SRCS})
 target_link_libraries(${PROJECT_NAME} 
     cpp_common
-    cpp_common_container
     cpp_common_inheritance
     gmock
     gtest
