+++ common/test/vector_test.cc
@@ -18,68 +18,86 @@ class TestObject {
     int mValue;
 };
 
-class VectorTest : public Test {};
+template <typename T>
+class VectorTest : public Test {
+public:
+    // void InitTest(int size = 0, TestObject val = TestObject()){
+    //     vec = T(size, val);
+    // }
+    T vec;
+};
 
-TEST_F(VectorTest, Resize) {
+using VectorTypes = ::testing::Types<container::vector<TestObject>, std::vector<TestObject>>;
+
+TYPED_TEST_SUITE(VectorTest, VectorTypes);
+TYPED_TEST(VectorTest, Resize) {
     {
-        std::vector<TestObject> stdvec;
-
         for (int i = 0; i < 6; ++i) {
-            stdvec.emplace_back();
+            this->vec.push_back(TestObject());
         }
-        std::cout << stdvec.capacity() << std::endl;
-        auto* t = &stdvec[5];
+        EXPECT_EQ(this->vec.capacity(),8);
+        auto* t = &this->vec[5];
         EXPECT_CALL(*t, Die()).Times(1);
-        stdvec.resize(5);
-        std::cout << stdvec.capacity() << std::endl;
+        this->vec.resize(5);
+        EXPECT_EQ(this->vec.capacity(), 8);
     }
 }
 
-TEST_F(VectorTest, Init) {
-    std::vector<TestObject> stdvecNoParamInit(5);
-    std::cout << stdvecNoParamInit.capacity() << std::endl;
+TYPED_TEST(VectorTest, Init) {
+    // this->InitTest(5);
+    // EXPECT_EQ(this->vec.capacity(), 10);
+    // std::vector<TestObject> stdvecNoParamInit(5);
+    // std::cout << stdvecNoParamInit.capacity() << std::endl;
 
-    std::vector<TestObject> stdvecParamInit(5, 1);
-    std::cout << stdvecParamInit.capacity() << std::endl;
+    // std::vector<TestObject> stdvecParamInit(5, 1);
+    // std::cout << stdvecParamInit.capacity() << std::endl;
 
-    std::vector<TestObject> stdvecVecInit(stdvecParamInit);
-    std::cout << stdvecParamInit.capacity() << std::endl;
+    // std::vector<TestObject> stdvecVecInit(stdvecParamInit);
+    // std::cout << stdvecParamInit.capacity() << std::endl;
 }
 
-TEST_F(VectorTest, FrontAndBack) {
-    std::vector<TestObject> stdvec;
-    auto& front_empty = stdvec.front();
-    auto& back_empty = stdvec.back();
-    // std::cout<<front_empty.mValue<<std::endl;
-    // std::cout<<back_empty.mValue<<std::endl;
-    stdvec.reserve(4);
-    auto& front = stdvec.front();
-    auto& back = stdvec.back();
+// TEST_F(VectorTest, FrontAndBack) {
+//     std::vector<TestObject> stdvec;
+//     auto& front_empty = stdvec.front();
+//     auto& back_empty = stdvec.back();
+//     // std::cout<<front_empty.mValue<<std::endl;
+//     // std::cout<<back_empty.mValue<<std::endl;
+//     stdvec.reserve(4);
+//     auto& front = stdvec.front();
+//     auto& back = stdvec.back();
+//     std::cout << stdvec.size() << std::endl;
+//     std::cout << stdvec[-1].mValue << std::endl;
+//     std::cout << stdvec[0].mValue << std::endl;
+//     std::cout << stdvec[1].mValue << std::endl;
+//     std::cout << stdvec[2].mValue << std::endl;
+//     std::cout << stdvec[3].mValue << std::endl;
 
-    EXPECT_EQ(front.mValue, stdvec[0].mValue);
-    EXPECT_EQ(back.mValue, stdvec[-1].mValue);
-}
+//     std::cout << front.mValue << std::endl;
+//     std::cout << back.mValue << std::endl;
+// }
 
-TEST_F(VectorTest, PushBack) {
-    std::vector<TestObject> stdvec;
-    stdvec.resize(4);
+// TEST_F(VectorTest, PushBack) {
+//     std::vector<TestObject> stdvec;
+//     stdvec.resize(5);
 
-    stdvec.push_back(TestObject());
-    EXPECT_EQ(stdvec.capacity(), 8);
-    stdvec.pop_back();
-    EXPECT_EQ(stdvec.capacity(), 8);
-}
+//     stdvec.push_back(TestObject());
+//     std::cout << stdvec.capacity() << std::endl;
+//     stdvec.pop_back();
+//     std::cout << stdvec.capacity() << std::endl;
+// }
 
-TEST_F(VectorTest, PopBack) {
-    std::vector<TestObject> stdvec;
-    stdvec.reserve(1);
-    EXPECT_DEATH(stdvec.pop_back(), "");
-}
+// TEST_F(VectorTest, PopBack) {
+//     std::vector<TestObject> stdvec;
+//     stdvec.reserve(1);
+//     stdvec.pop_back();
+//     // stdvec.pop_back();
+// }
 
-TEST_F(VectorTest, Clear) {
-    std::vector<TestObject> stdvec(5);
-    EXPECT_EQ(stdvec.capacity(), 5);
-    stdvec.clear();
-    EXPECT_EQ(stdvec.capacity(), 5);
-}
+// TEST_F(VectorTest, Clear) {
+//     std::vector<TestObject> stdvec(5);
+//     EXPECT_EQ(stdvec.capacity(), 5);
+//     stdvec.clear();
+//     EXPECT_EQ(stdvec.capacity(), 5);
+// }
+
 }  // namespace cpp::common::test
